Refactored the project, ported over to crossterm instead of pancurses because it behaves better in a Windows environment.

+++ src/event.rs
@@ -1,6 +1,3 @@
-// This provides a type-safe way to handle input. Instead of dealing with raw pancurses input codes,
-// you get nicely-typed events. The get_input() method in Window translates pancurses' input into
-// our event type.
 #[derive(Debug, Clone, PartialEq)]
 pub enum Event {
     Character(char),
@@ -11,6 +8,6 @@ pub enum Event {
     Delete,
     Backspace,
     Enter,
-    FunctionKey(i32),
+    FunctionKey(u8),
     Unknown,
 }
